Add Flyway lib and implement it, add sql schemas for migration, fixed bug with not found sql_sequence, remove trigger from DB, add oracl lib to maven build

+++ epam-chadov-test-task1/src/main/java/com/epam/chadov/task1/database/GenericDao.java
@@ -17,5 +17,5 @@ public interface GenericDao<T extends AbstractEntity> {
 
     boolean deleteNews(News news);
 
-    T getById(Long id);
+    T getById(Integer id);
 }
